+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # 14.1.0
 
+- Add `modulation::Fourier`
+  - Fix [#110](https://github.com/shinolab/autd3/issues/110): Multi-frequency sine modulation? 
+
+# 14.1.0
+
 - Fix [#93](https://github.com/shinolab/autd3/issues/93): pyautd3 package contains unnecessary dynamic libraries
 - Fix [#94](https://github.com/shinolab/autd3/issues/94): pyautd3 library should clarify its dependence on numpy
 - Fix [#108](https://github.com/shinolab/autd3/issues/108): OsalTimer on macOS causes segmentation fault